Implement an "AudioStart" event.
Because the current implementation of the state machine ignores transitions to the
same state, we need to change states to actually do anything. When we change to the
next or previous message, we need to move to a different state than "NetMessage", so
that we can go to the (yet) next message before the current message stops. The
"AudioStart" event lets us implement a "Playing" state, from which we can navigate.